+++ config/settings/test.py
@@ -30,3 +30,4 @@ EMAIL_BACKEND = "django.core.mail.backends.locmem.EmailBackend"
 TEMPLATES[0]["OPTIONS"]["debug"] = True  # type: ignore # noqa F405
 # Your stuff...
 # ------------------------------------------------------------------------------
+DOMAIN_URL = "http://127.0.0.1:8000"