+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2023.03.07
+- Add base tests for model, views, serializers etc.
+- Allow Create shorten urls for not logged user
+
 ## 2023.03.06
 - Base Application engine with user's management
 - tinyurls App with API Endpoints: